requirements.txt
Django>=4.2
celery>=5.3
redis>=5.0

django-celery-beat
django-celery-results

django-cors-headers
djangorestframework
django-filter
djangorestframework-simplejwt

psycopg[binary]
